+++ AddyStatisticWidget/AddyStatisticWidget.swift
@@ -265,15 +265,15 @@ struct AliasWidgetRowView: View {
                                 aliasDescription =  String(format: String(localized: "s_s_s"),
                                                            description,
                                                            String(format: NSLocalizedString("created_at_s", comment: ""),
-                                                                  DateTimeUtils.turnStringIntoLocalString(alias.created_at)),
+                                                                  DateTimeUtils.convertStringToLocalTimeZoneString(alias.created_at)),
                                                            String(format: String(localized: "updated_at_s"),
-                                                                  DateTimeUtils.turnStringIntoLocalString(alias.updated_at)))
+                                                                  DateTimeUtils.convertStringToLocalTimeZoneString(alias.updated_at)))
                             } else {
                                 aliasDescription =  String(format: String(localized: "s_s"),
                                                            String(format: NSLocalizedString("created_at_s", comment: ""),
-                                                                  DateTimeUtils.turnStringIntoLocalString(alias.created_at)),
+                                                                  DateTimeUtils.convertStringToLocalTimeZoneString(alias.created_at)),
                                                            String(format: String(localized: "updated_at_s"),
-                                                                  DateTimeUtils.turnStringIntoLocalString(alias.updated_at)))
+                                                                  DateTimeUtils.convertStringToLocalTimeZoneString(alias.updated_at)))
                             }
                         }
                 }