+++ addy/View/mainview/MainView.swift
@@ -13,13 +13,13 @@ import LocalAuthentication
 struct MainView: View {
     @EnvironmentObject var mainViewState: MainViewState
     @StateObject private var aliasesViewState = AliasesViewState.shared // Needs to be shared so that filters can be applied from other views
-
+    
     // MARK: Share sheet AND MailTo tap action
     @State var pendingURLFromShareViewController: IdentifiableURL? = nil
     // MARK: END Share sheet AND MailTo tap action
     
     @Environment(\.scenePhase) var scenePhase
-
+    
     
     @State private var apiTokenExpiryText = ""
     @State private var subscriptionExpiryText = ""
@@ -71,10 +71,12 @@ struct MainView: View {
                             SettingsManager(encrypted: false).putSettingsInt(key: .timesTheAppHasBeenOpened, int:
                                                                                 SettingsManager(encrypted: false).getSettingsInt(key: .timesTheAppHasBeenOpened) + 1)
                             
-                            #if DEBUG
+#if DEBUG
                             print("App has been opened \(SettingsManager(encrypted: false).getSettingsInt(key: .timesTheAppHasBeenOpened)) times")
-                            #endif
-
+#endif
+                            
+                            
+                            openDefaultPage()
                         })
                         .alert(isPresented: Binding<Bool>(
                             get: { self.mainViewState.showApiExpirationWarning || self.mainViewState.showSubscriptionExpirationWarning },
@@ -188,19 +190,8 @@ struct MainView: View {
                         mainViewState.selectedTab = .aliases
                     }
                     
+                    checkForPendingURLFromShareViewController()
                     
-                    // Check if there are pendingURLFromShareViewController
-                    if let url = SettingsManager(encrypted: true).getSettingsString(key: .pendingURLFromShareViewController) {
-                        
-                        if url.starts(with: "addyio://") {
-                            UIApplication.shared.open(URL(string: "\(url)")!, options: [:], completionHandler: nil)
-                        } else {
-                            pendingURLFromShareViewController = IdentifiableURL(url: URL(string: "\(url)")!)
-                        }
-                        
-                        // Remove to prevent any future references
-                        SettingsManager(encrypted: true).removeSetting(key: .pendingURLFromShareViewController)
-                    }
                     
                     // Check this every time the app is come to foreground
                     checkForAlerts()
@@ -263,6 +254,29 @@ struct MainView: View {
         
     }
     
+    private func openDefaultPage(){
+        // Check if the value exists in the array, default (but dont reset) to home if not (this could occur if eg. a tablet backup (which has more options) gets restored on mobile)
+        // Don't reset the value as this app could be opened in splitscreen, we don't want to reset the value then.
+        
+        let destinations = horizontalSize == .regular ? Destination.otherCases : Destination.iPhoneCases
+        let startupPage = SettingsManager(encrypted: false).getSettingsString(key: .startupPage) ?? "home"
+        
+        if destinations.contains(where: { $0.value == startupPage }){
+            switch startupPage {
+            case Destination.home.value: mainViewState.selectedTab = .home
+            case Destination.aliases.value: mainViewState.selectedTab =  .aliases
+            case Destination.recipients.value: mainViewState.selectedTab =  .recipients
+            case Destination.usernames.value: mainViewState.selectedTab =  .usernames
+            case Destination.domains.value: mainViewState.selectedTab =  .domains
+            case Destination.failedDeliveries.value: mainViewState.selectedTab =  .failedDeliveries
+            case Destination.rules.value: mainViewState.selectedTab =  .rules
+            case Destination.settings.value: mainViewState.selectedTab =  .settings
+            default:
+                break
+            }
+        }
+    }
+    
     
     private func handleURL(url: URL) {
         switch url.host {
@@ -282,7 +296,7 @@ struct MainView: View {
             do {
                 let user = try await NetworkHelper().getUserResource()
                 if let subscriptionEndsAt = user?.subscription_ends_at {
-                    let expiryDate = try DateTimeUtils.turnStringIntoLocalDateTime(subscriptionEndsAt) // Get the expiry date
+                    let expiryDate = try DateTimeUtils.convertStringToLocalTimeZoneDate(subscriptionEndsAt) // Get the expiry date
                     let currentDateTime = Date() // Get the current date
                     let deadLineDate = Calendar.current.date(byAdding: .day, value: -7, to: expiryDate) // Subtract 7 days from the expiry date
                     if let deadLineDate = deadLineDate, currentDateTime > deadLineDate {
@@ -304,12 +318,27 @@ struct MainView: View {
         }
     }
     
+    private func checkForPendingURLFromShareViewController(){
+        // Check if there are pendingURLFromShareViewController
+        if let url = SettingsManager(encrypted: true).getSettingsString(key: .pendingURLFromShareViewController) {
+            
+            // eg. addyio://alias/\(aliasId)" (from ShareViewController)
+            if url.starts(with: "addyio://") {
+                UIApplication.shared.open(URL(string: "\(url)")!, options: [:], completionHandler: nil)
+            } else {
+                pendingURLFromShareViewController = IdentifiableURL(url: URL(string: "\(url)")!)
+            }
+            
+            // Remove to prevent any future references
+            SettingsManager(encrypted: true).removeSetting(key: .pendingURLFromShareViewController)
+        }
+    }
     
     private func checkTokenExpiry() async {
         do {
             let apiTokenDetails = try await NetworkHelper().getApiTokenDetails()
             if let expiresAt = apiTokenDetails?.expires_at {
-                let expiryDate = try DateTimeUtils.turnStringIntoLocalDateTime(expiresAt) // Get the expiry date
+                let expiryDate = try DateTimeUtils.convertStringToLocalTimeZoneDate(expiresAt) // Get the expiry date
                 let currentDateTime = Date() // Get the current date
                 let deadLineDate = Calendar.current.date(byAdding: .day, value: -5, to: expiryDate) // Subtract 5 days from the expiry date
                 if let deadLineDate = deadLineDate, currentDateTime > deadLineDate {
@@ -500,7 +529,7 @@ struct MainView: View {
     }
 }
 
-enum Destination: Hashable, CaseIterable {
+public enum Destination: Hashable, CaseIterable {
     case home, aliases, recipients, usernames, domains, failedDeliveries, rules, settings, subscription
     
     
@@ -523,6 +552,22 @@ enum Destination: Hashable, CaseIterable {
         case .rules: return "rules"
         case .settings: return "settings"
         case .subscription: return "subscription"
+            
+        }
+    }
+    
+    var value: String {
+        switch self {
+        case .home: return "home"
+        case .aliases: return "aliases"
+        case .recipients: return "recipients"
+        case .usernames: return "usernames"
+        case .domains: return "domains"
+        case .failedDeliveries: return "failed_deliveries"
+        case .rules: return "rules"
+        case .settings: return "settings"
+        case .subscription: return "subscription"
+            
         }
     }
     
@@ -537,6 +582,7 @@ enum Destination: Hashable, CaseIterable {
         case .rules: return "checklist"
         case .settings: return "gear"
         case .subscription: return "creditcard.fill"
+            
         }
     }
     
@@ -564,10 +610,8 @@ enum Destination: Hashable, CaseIterable {
             refreshGeneralData?()
         }))
         case .settings: return AnyView(AppSettingsView(horizontalSize: horizontalSize))
-        case .subscription: return AnyView(ManageSubscriptionView(horizontalSize: horizontalSize, shouldHideNavigationBarBackButtonSubscriptionView: .constant(false)))
-            
+        case .subscription: return AnyView(ManageSubscriptionView(horizontalSize: horizontalSize, shouldHideNavigationBarBackButtonSubscriptionView: .constant(false))) // This actually never gets called, it is required because the destination need to contain the subscription destination for checkForAnyInteractiveActions in ProfileBottomSheet.swift
         }
-        
     }
 }