+++ addy_shared/addy_shared/managers/SettingsManager.swift
@@ -41,6 +41,7 @@ public class SettingsManager {
         case userResourceExtended
         case backgroundServiceCacheMostActiveAliasesData
         case timesTheAppHasBeenOpened
+        case startupPage
 
         
         // WearOS